+++ DataCenter/Raw/Services/Items/RawItemsService.cs
@@ -6,8 +6,8 @@ namespace DBI.DataCenter.Raw.Services.Items;
 /// </summary>
 public class RawItemsService(IReadOnlyCollection<RawItem> items)
 {
-    readonly Dictionary<ushort, RawItem> _items = items.ToDictionary(item => item.Id, item => item);
+    readonly Dictionary<int, RawItem> _items = items.ToDictionary(item => (int)item.Id, item => item);
 
-    public RawItem? GetItem(ushort itemId) => _items.GetValueOrDefault(itemId);
+    public RawItem? GetItem(int itemId) => _items.GetValueOrDefault(itemId);
     public IEnumerable<RawItem> GetItems() => _items.Values;
 }
